feat: Improve project structure and fix some margins
