+++ src/transformations/transformFormat.ts
@@ -1,5 +1,8 @@
+import { renameFields, skipFields } from './transformData';
+
 import { TransformableResource } from '@/schemas/configuration-schema';
 import { csv2json } from 'json-2-csv';
+import { narrowType } from './transformType';
 
 export function transformData(resource: TransformableResource, data: unknown) {
     let transformedData = transformType(resource, data);
@@ -85,64 +88,4 @@ function isTabularJson(json: unknown): json is TabularJson {
 
 function transformCsvData(csv: unknown) {
     return csv2json(String(csv)) as Record<string, never>[];
-}
-
-function skipFields(records: Record<string, never>[], skipFieldsRegex: string) {
-    if (records.length === 0) {
-        return records;
-    }
-
-    const regex = new RegExp(skipFieldsRegex, 'u');
-    const fieldsToRemove = Object.keys(records[0]).filter((key) => regex.test(key));
-
-    if (fieldsToRemove.length === 0) {
-        return records;
-    }
-
-    return records.map((record) => {
-        const filteredFields = Object.entries(record).filter(([key, _]) => !fieldsToRemove.includes(key));
-        return Object.fromEntries(filteredFields);
-    });
-}
-
-function renameFields(records: Record<string, never>[], renameFieldsObj: Record<string, string>) {
-    return records.map((record) => {
-        const renamedObj = {} as Record<string, never>;
-        Object.entries(record).forEach(([key, value]) => {
-            const renamedKey = renameFieldsObj[key] ?? key;
-            renamedObj[renamedKey] = value;
-        });
-        return renamedObj;
-    });
-}
-
-function narrowType(records: Record<string, never>[], resource: TransformableResource) {
-    return records.map(
-        (record) =>
-            Object.fromEntries(
-                Object.entries(record).map(([key, value]) => {
-                    const stringValue = String(value).toLowerCase();
-                    if (stringValue === 'true') {
-                        return [key, true];
-                    }
-                    if (stringValue === 'false') {
-                        return [key, false];
-                    }
-                    const parsedValue =
-                        resource.numberFormat === 'en'
-                            ? Number(stringValue)
-                            : parseGermanNumberToInternationalFormat(stringValue);
-                    if (!Number.isNaN(parsedValue)) {
-                        return [key, parsedValue];
-                    }
-                    return [key, value];
-                }),
-            ) as Record<string, string | number | boolean>,
-    );
-}
-
-function parseGermanNumberToInternationalFormat(value: string) {
-    const parsedValue = value.replace('.', '').replace(',', '.');
-    const isNumeric = !Number.isNaN(Number(parsedValue));
-    return isNumeric ? Number(parsedValue) : Number.NaN;
 }
